+++ src/Seek.hs
@@ -7,20 +7,42 @@
 module Seek where
 
 import Control.Monad
+import Data.Maybe
+import Data.Int
+import Data.Word
 import Data.IORef
 import Data.GI.Base.Signals
-import GI.GLib
-import GI.GObject
+import qualified GI.GLib
+import qualified GI.GObject
 import qualified GI.Gtk
 import qualified GI.Gst
 
 import qualified Records as R
-import Playbin
 import Utils
 
 addSeekHandlers :: R.Application -> IO ()
 addSeekHandlers
-  R.Application
+  application@R.Application
+    { R.guiObjects =
+        R.GuiObjects
+          { R.seekScale = seekScale
+          }
+    , R.playbin = playbin
+    }
+  = do
+  seekScaleHandlerId <-
+    GI.Gtk.onRangeValueChanged seekScale $
+      onRangeValueChanged application
+  void $
+    GI.GLib.timeoutAdd GI.GLib.PRIORITY_DEFAULT 41 $
+      updateSeekScale
+        playbin
+        seekScale
+        seekScaleHandlerId
+
+onRangeValueChanged :: R.Application -> IO ()
+onRangeValueChanged
+  application@R.Application
     { R.guiObjects =
         R.GuiObjects
           { R.seekScale = seekScale
@@ -32,25 +54,6 @@ addSeekHandlers
     , R.playbin = playbin
     }
   = do
-  seekScaleHandlerId <-
-    GI.Gtk.onRangeValueChanged seekScale $
-      onRangeValueChanged
-        playbin
-        seekScale
-        videoInfoRef
-  void $
-    GI.GLib.timeoutAdd GI.GLib.PRIORITY_DEFAULT 41 $
-      updateSeekScale
-        playbin
-        seekScale
-        seekScaleHandlerId
-
-onRangeValueChanged :: GI.Gst.Element -> GI.Gtk.Scale -> IORef R.VideoInfo -> IO ()
-onRangeValueChanged
-  playbin
-  seekScale
-  videoInfoRef
-  = do
   maybeDurationAndPosition <- queryPlaybinForDurationAndPosition playbin
   case maybeDurationAndPosition of
     (Just (duration, _)) -> do
@@ -59,13 +62,13 @@ onRangeValueChanged
             clamp 0 duration $
                 doubleToInt64 $
                   int64toDouble duration * percentage
-      void $ GI.Gst.elementSeekSimple playbin GI.Gst.FormatTime [ GI.Gst.SeekFlagsFlush ] position
-      videoInfo <- readIORef videoInfoRef
+      seekTo application (Just position) Nothing
       -- When playing over a network (not local),
       -- sending too many asynchronous seek requests will cause GStreamer to error.
       -- Block until this seek request finishes which will block the main GTK main thread.
       -- Note that by blocking here, the seek scale will not slide.
       -- The user will have to point and click on the scale track.
+      videoInfo <- readIORef videoInfoRef
       unless (R.isLocalFile videoInfo) $
         void $
           GI.Gst.elementGetState playbin GI.Gst.CLOCK_TIME_NONE
@@ -99,3 +102,62 @@ hideOrShowSeekScale seekScale True = do
 hideOrShowSeekScale seekScale False = do
   GI.Gtk.widgetSetSensitive seekScale False
   GI.Gtk.widgetHide seekScale
+
+seekTo :: R.Application -> Maybe Int64 -> Maybe Word32 -> IO ()
+seekTo
+  R.Application
+    { R.guiObjects =
+        R.GuiObjects
+          { R.videoSpeedSelectionComboboxText = videoSpeedSelectionComboboxText
+          }
+    , R.ioRefs =
+        R.IORefs
+          { R.videoInfoRef = videoInfoRef
+          }
+    , R.playbin = playbin
+    }
+  maybeDesiredPosition
+  maybeDelay
+  = do
+  videoInfo      <- readIORef videoInfoRef
+  activeId       <- GI.Gtk.comboBoxGetActive videoSpeedSelectionComboboxText
+  let isSeekable = R.isSeekable videoInfo
+  let isVideo    = R.isVideo videoInfo
+  case (isVideo, isSeekable, activeId) of
+    (True, True, 0) -> seekToWithPlaybackRate 0.5
+    (True, True, 1) -> seekToWithPlaybackRate 1.0
+    (True, True, 2) -> seekToWithPlaybackRate 1.5
+    _ -> return ()
+  where
+    seekToWithPlaybackRate :: Double -> IO ()
+    seekToWithPlaybackRate rate =
+      case maybeDelay of
+        Just delay ->
+          void $
+            GI.GLib.timeoutAdd
+              GI.GLib.PRIORITY_DEFAULT
+              (if delay < 0 then 0 else delay) $ do
+                createAndSendSeekEvent
+                return False
+        Nothing -> createAndSendSeekEvent
+      where
+        createAndSendSeekEvent :: IO ()
+        createAndSendSeekEvent = do
+          maybeDurationPosition <- queryPlaybinForDurationAndPosition playbin
+          case maybeDurationPosition of
+            (Just (_, currentPosition)) -> do
+              let position = fromMaybe currentPosition maybeDesiredPosition
+              let flags =
+                      GI.Gst.SeekFlagsFlush
+                    : [GI.Gst.SeekFlagsTrickmode | rate /= 1.0]
+              void $
+                GI.Gst.elementSeek
+                  playbin
+                  rate
+                  GI.Gst.FormatTime
+                  flags
+                  GI.Gst.SeekTypeSet
+                  position
+                  GI.Gst.SeekTypeNone
+                  0
+            Nothing -> return ()