+++ src/Window.hs
@@ -25,6 +25,8 @@ import Reset
 import Mouse
 import PlayPause
 import Seek
+import BottomControlsBoxVisibility
+import Utils
 
 addWindowHandlers :: R.Application -> [R.Application -> IO ()] -> IO ()
 addWindowHandlers
@@ -43,16 +45,23 @@ addWindowHandlers
   onWidgetRealizeCallbacks
   = do
   void $
-    GI.Gtk.onWidgetRealize videoWidget $
-      windowRealizedHandler
-        application
-        onWidgetRealizeCallbacks
+    GI.Gtk.onWidgetRealize
+      videoWidget $
+        windowRealizedHandler
+          application
+          onWidgetRealizeCallbacks
   void $
-    GI.Gtk.onWidgetWindowStateEvent window (widgetWindowStateEventHandler isWindowFullScreenRef)
+    GI.Gtk.onWidgetWindowStateEvent
+      window $
+        widgetWindowStateEventHandler isWindowFullScreenRef
   void $
-    GI.Gtk.onWidgetDestroy window (windowDestroyHandler playbin)
+    GI.Gtk.onWidgetDestroy
+      window $
+        windowDestroyHandler playbin
   void $
-    GI.GLib.timeoutAddSeconds GI.GLib.PRIORITY_DEFAULT 1 (adjustWindow application)
+    GI.GLib.timeoutAddSeconds
+      GI.GLib.PRIORITY_DEFAULT 1 $
+        adjustWindowPeriodically application
 
 windowRealizedHandler
   ::  R.Application
@@ -62,15 +71,17 @@ windowRealizedHandler
   application@R.Application
     { R.guiObjects =
         guiObjects@R.GuiObjects
-          { R.videoWidget = videoWidget
-          , R.seekScale   = seekScale
+          { R.fileChooserButton = fileChooserButton
+          , R.videoWidget       = videoWidget
+          , R.seekScale         = seekScale
           }
     }
   onWidgetRealizeCallbacks
   = do
   let eventMask = [GI.Gdk.EventMaskAllEventsMask]
-  GI.Gtk.widgetAddEvents videoWidget eventMask
-  GI.Gtk.widgetAddEvents seekScale   eventMask
+  GI.Gtk.widgetAddEvents fileChooserButton eventMask
+  GI.Gtk.widgetAddEvents videoWidget       eventMask
+  GI.Gtk.widgetAddEvents seekScale         eventMask
   resetGuiObjects guiObjects
   mapM_ (\ f -> f application) onWidgetRealizeCallbacks
 
@@ -91,14 +102,13 @@ windowDestroyHandler playbin = do
   _ <- GI.Gst.objectUnref playbin
   GI.Gtk.mainQuit
 
-adjustWindow :: R.Application -> IO Bool
-adjustWindow
+adjustWindowPeriodically :: R.Application -> IO Bool
+adjustWindowPeriodically
   application@R.Application
     { R.guiObjects =
         R.GuiObjects
-          { R.window               = window
-          , R.fileChooserButton    = fileChooserButton
-          , R.bottomControlsGtkBox = bottomControlsGtkBox
+          { R.window         = window
+          , R.topControlsBox = topControlsBox
           }
     , R.ioRefs =
         R.IORefs
@@ -112,16 +122,13 @@ adjustWindow
   mouseMovedLast       <- readIORef mouseMovedLastRef
   timeNow              <- fmap round getPOSIXTime
   (_, playBinState, _) <- GI.Gst.elementGetState playbin (fromIntegral GI.Gst.MSECOND :: GHC.Word.Word64)
-  let isPlaying = GI.Gst.StatePlaying == playBinState
-  let delta     = timeNow - mouseMovedLast
-  let isVideo   = R.isVideo videoInfo
+  let isPlaying        = GI.Gst.StatePlaying == playBinState
+  let delta            = timeNow - mouseMovedLast
+  let isVideo          = R.isVideo videoInfo
   when (isPlaying && isVideo && delta >= hideOnScreenControlsInterval) $ do
-    GI.Gtk.widgetHide fileChooserButton
-    GI.Gtk.widgetHide bottomControlsGtkBox
-    styleContext <- GI.Gtk.widgetGetStyleContext bottomControlsGtkBox
-    GI.Gtk.styleContextRemoveClass styleContext "movie-monad-fade-in"
+    GI.Gtk.widgetHide topControlsBox
     setCursor window (Just "none")
-    atomicWriteIORef mouseMovedLastRef timeNow
+    void $ fadeOutBottomControlsBox application
   fillWindowWithVideo application
   return True
 
@@ -153,7 +160,7 @@ calculateWindowSize :: R.GuiObjects -> Int -> R.VideoInfo -> IO (Maybe (Int32, I
 calculateWindowSize _ _ R.VideoInfo { R.isVideo = False } = return Nothing
 calculateWindowSize
   R.GuiObjects
-    { R.fileChooserButton = fileChooserButton
+    { R.topControlsBox = topControlsBox
     }
   desiredWidth
   R.VideoInfo
@@ -161,18 +168,18 @@ calculateWindowSize
     , R.videoHeight = videoHeight
     }
   = do
-  fileChooserButtonIsVisible <- GI.Gtk.widgetGetVisible fileChooserButton
-  fileChooserButtonHeight    <- GI.Gtk.widgetGetAllocation fileChooserButton >>= GI.Gdk.getRectangleHeight
-  let videoWidthDouble  = fromIntegral videoWidth  :: Double
-  let videoHeightDouble = fromIntegral videoHeight :: Double
+  topControlsBoxIsVisible <- GI.Gtk.widgetGetVisible topControlsBox
+  topControlsBoxHeight    <- GI.Gtk.widgetGetAllocation topControlsBox >>= GI.Gdk.getRectangleHeight
+  let videoWidthDouble       = fromIntegral videoWidth  :: Double
+  let videoHeightDouble      = fromIntegral videoHeight :: Double
   let ratio =
         if videoWidthDouble <= 0.0
           then 0.0
           else videoHeightDouble / videoWidthDouble
   let desiredWidthDouble = fromIntegral desiredWidth :: Double
   let topMargin =
-        if fileChooserButtonIsVisible
-          then (fromIntegral fileChooserButtonHeight :: Double)
+        if topControlsBoxIsVisible
+          then (fromIntegral topControlsBoxHeight :: Double)
           else 0.0
   let bottomMargin = 0.0
   let height = topMargin + (desiredWidthDouble * ratio) + bottomMargin
@@ -188,32 +195,53 @@ setWindowSize
   =
   GI.Gtk.windowResize window width (if height <= 0 then 1 else height)
 
-setupWindowForPlayback :: R.GuiObjects -> Bool -> Bool -> Int32 -> Int32 -> IO ()
+setupWindowForPlayback :: R.Application -> Bool -> Bool -> Int32 -> Int32 -> IO ()
 setupWindowForPlayback
-  guiObjects@R.GuiObjects
-    { R.window               = window
-    , R.videoWidget          = videoWidget
-    , R.bottomControlsGtkBox = bottomControlsGtkBox
-    , R.seekScale            = seekScale
-    , R.playPauseButton      = playPauseButton
-    , R.repeatCheckButton    = repeatCheckButton
-    , R.fullscreenButton     = fullscreenButton
-    , R.playImage            = playImage
-    , R.pauseImage           = pauseImage
+  application@R.Application
+    { R.guiObjects =
+        guiObjects@R.GuiObjects
+          { R.window                          = window
+          , R.videoWidget                     = videoWidget
+          , R.topControlsBox                  = topControlsBox
+          , R.bottomControlsBox               = bottomControlsBox
+          , R.seekScale                       = seekScale
+          , R.playPauseButton                 = playPauseButton
+          , R.repeatCheckButton               = repeatCheckButton
+          , R.fullscreenButton                = fullscreenButton
+          , R.videoSpeedSelectionComboboxText = videoSpeedSelectionComboboxText
+          , R.subtitleSelectionComboBoxText   = subtitleSelectionComboBoxText
+          , R.playImage                       = playImage
+          , R.pauseImage                      = pauseImage
+          }
     }
   isSeekable
   isWindowFullScreen
   windowWidth
   windowHeight
   = do
-  GI.Gtk.widgetShow bottomControlsGtkBox
   GI.Gtk.widgetShow videoWidget
   GI.Gtk.widgetShow playPauseButton
   GI.Gtk.widgetShow repeatCheckButton
   GI.Gtk.widgetShow fullscreenButton
   GI.Gtk.setToggleButtonActive repeatCheckButton False
-  GI.Gtk.widgetSetSizeRequest window windowMinimumSize 100
+  windowMinimumHeight <- getWindowMinimumHeight
+  GI.Gtk.widgetSetSizeRequest window windowMinimumWidth windowMinimumHeight
+  GI.Gtk.comboBoxSetActive videoSpeedSelectionComboboxText 1
+  GI.Gtk.comboBoxSetActive subtitleSelectionComboBoxText 0
   setPlayPauseButton playPauseButton playImage pauseImage True
   unless isWindowFullScreen $
     setWindowSize guiObjects windowWidth windowHeight
   hideOrShowSeekScale seekScale isSeekable
+  void $ fadeInBottomControlsBox application 0
+  where
+    getWindowMinimumHeight :: IO Int32
+    getWindowMinimumHeight = do
+      (_, topControlsboxHeight)     <- getWidgetSize topControlsBox
+      (_, bottomControlsBoxHeight)  <- getWidgetSize bottomControlsBox
+      bottomControlsBoxMarginTop    <- GI.Gtk.widgetGetMarginTop bottomControlsBox
+      bottomControlsBoxMarginBottom <- GI.Gtk.widgetGetMarginBottom bottomControlsBox
+      return $
+          topControlsboxHeight
+        + bottomControlsBoxHeight
+        + bottomControlsBoxMarginTop
+        + bottomControlsBoxMarginBottom